Remove unused backup UI, show full dashboard and enable profile route
