--- frontend/_backup_ui/src/context/SidebarContext.jsx
@@ -1,41 +1,0 @@
-import  { createContext, useContext, useState, useEffect } from 'react';
-import { useLocation } from 'react-router-dom';
-
-const SidebarContext = createContext();
-
-export const useSidebar = () => useContext(SidebarContext);
-
-export const SidebarProvider = ({ children }) => {
-  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
-  const [isMobile, setIsMobile] = useState(window.innerWidth <= 640);
-  const location = useLocation();
-
-  useEffect(() => {
-    const handleResize = () => setIsMobile(window.innerWidth <= 640);
-    window.addEventListener('resize', handleResize);
-    return () => window.removeEventListener('resize', handleResize);
-  }, []);
-
-  useEffect(() => {
-    if (isMobile) setIsSidebarOpen(false);
-  }, [location.pathname, isMobile]);
-
-  const layoutDirection = isMobile
-    ? isSidebarOpen
-      ? 'flex-col-reverse'
-      : 'flex-col'
-    : 'flex-row-reverse';
-
-  return (
-    <SidebarContext.Provider
-      value={{
-        isSidebarOpen,
-        setIsSidebarOpen,
-        isMobile,
-        layoutDirection,
-      }}
-    >
-      {children}
-    </SidebarContext.Provider>
-  );
-};
