--- frontend/_backup_ui/tailwind.config.js
@@ -1,195 +1,0 @@
-import plugin from "tailwindcss/plugin";
-
-export default {
-  darkMode: ["class"],
-  content: [
-    "./pages/**/*.{ts,tsx,js,jsx}",
-    "./components/**/*.{ts,tsx,js,jsx}",
-    "./app/**/*.{ts,tsx,js,jsx}",
-    "./src/**/*.{ts,tsx,js,jsx}",
-  ],
-  prefix: "",
-  future: {
-    hoverOnlyWhenSupported: true,
-  },
-  theme: {
-    container: {
-      center: true,
-      padding: '2rem',
-      screens: {
-        '2xl': '1400px'
-      }
-    },
-    extend: {
-      colors: {
-        border: 'hsl(var(--border))',
-        input: 'hsl(var(--input))',
-        ring: 'hsl(var(--ring))',
-        background: 'hsl(var(--background))',
-        foreground: 'hsl(var(--foreground))',
-        primary: {
-          DEFAULT: 'hsl(var(--primary))',
-          foreground: 'hsl(var(--primary-foreground))',
-          hover: 'hsl(var(--primary-hover))',
-          light: 'hsl(var(--primary-light))'
-        },
-        secondary: {
-          DEFAULT: 'hsl(var(--secondary))',
-          foreground: 'hsl(var(--secondary-foreground))',
-          hover: 'hsl(var(--secondary-hover))'
-        },
-        destructive: {
-          DEFAULT: 'hsl(var(--destructive))',
-          foreground: 'hsl(var(--destructive-foreground))'
-        },
-        success: {
-          DEFAULT: 'hsl(var(--success))',
-          foreground: 'hsl(var(--success-foreground))'
-        },
-        warning: {
-          DEFAULT: 'hsl(var(--warning))',
-          foreground: 'hsl(var(--warning-foreground))'
-        },
-        muted: {
-          DEFAULT: 'hsl(var(--muted))',
-          foreground: 'hsl(var(--muted-foreground))'
-        },
-        accent: {
-          DEFAULT: 'hsl(var(--accent))',
-          foreground: 'hsl(var(--accent-foreground))',
-          hover: 'hsl(var(--accent-hover))'
-        },
-        popover: {
-          DEFAULT: 'hsl(var(--popover))',
-          foreground: 'hsl(var(--popover-foreground))'
-        },
-        card: {
-          DEFAULT: 'hsl(var(--card))',
-          foreground: 'hsl(var(--card-foreground))',
-          hover: 'hsl(var(--card-hover))'
-        },
-        sidebar: {
-          DEFAULT: 'hsl(var(--sidebar-background))',
-          foreground: 'hsl(var(--sidebar-foreground))',
-          primary: 'hsl(var(--sidebar-primary))',
-          'primary-foreground': 'hsl(var(--sidebar-primary-foreground))',
-          accent: 'hsl(var(--sidebar-accent))',
-          'accent-foreground': 'hsl(var(--sidebar-accent-foreground))',
-          border: 'hsl(var(--sidebar-border))',
-          ring: 'hsl(var(--sidebar-ring))'
-        }
-      },
-      fontFamily: {
-        heading: ['var(--font-heading)'],
-        body: ['var(--font-body)']
-      },
-      backgroundImage: {
-        'gradient-primary': 'var(--gradient-primary)',
-        'gradient-secondary': 'var(--gradient-secondary)',
-        'gradient-hero': 'var(--gradient-hero)',
-        'gradient-card': 'var(--gradient-card)'
-      },
-      boxShadow: {
-        'sm': 'var(--shadow-sm)',
-        'md': 'var(--shadow-md)',
-        'lg': 'var(--shadow-lg)',
-        'xl': 'var(--shadow-xl)',
-        'glow': 'var(--shadow-glow)'
-      },
-      borderRadius: {
-        lg: 'var(--radius)',
-        md: 'calc(var(--radius) - 2px)',
-        sm: 'calc(var(--radius) - 4px)'
-      },
-      keyframes: {
-        'accordion-down': {
-          from: { height: '0' },
-          to: { height: 'var(--radix-accordion-content-height)' }
-        },
-        'accordion-up': {
-          from: { height: 'var(--radix-accordion-content-height)' },
-          to: { height: '0' }
-        },
-        'fade-in': {
-          '0%': { opacity: '0', transform: 'translateY(20px)' },
-          '100%': { opacity: '1', transform: 'translateY(0)' }
-        },
-        'fade-in-up': {
-          '0%': { opacity: '0', transform: 'translateY(40px)' },
-          '100%': { opacity: '1', transform: 'translateY(0)' }
-        },
-        'scale-in': {
-          '0%': { opacity: '0', transform: 'scale(0.95)' },
-          '100%': { opacity: '1', transform: 'scale(1)' }
-        },
-        'slide-in-right': {
-          '0%': { opacity: '0', transform: 'translateX(100%)' },
-          '100%': { opacity: '1', transform: 'translateX(0)' }
-        },
-        'slide-in-left': {
-          '0%': { opacity: '0', transform: 'translateX(-100%)' },
-          '100%': { opacity: '1', transform: 'translateX(0)' }
-        },
-        'bounce-in': {
-          '0%': { opacity: '0', transform: 'scale(0.3)' },
-          '50%': { opacity: '1', transform: 'scale(1.05)' },
-          '70%': { transform: 'scale(0.9)' },
-          '100%': { opacity: '1', transform: 'scale(1)' }
-        },
-        'glow': {
-          '0%, 100%': { boxShadow: '0 0 20px hsla(214, 85%, 65%, 0.3)' },
-          '50%': { boxShadow: '0 0 40px hsla(214, 85%, 65%, 0.6)' }
-        }
-      },
-      animation: {
-        'accordion-down': 'accordion-down 0.2s ease-out',
-        'accordion-up': 'accordion-up 0.2s ease-out',
-        'fade-in': 'fade-in 0.6s ease-out',
-        'fade-in-up': 'fade-in-up 0.8s ease-out',
-        'scale-in': 'scale-in 0.4s ease-out',
-        'slide-in-right': 'slide-in-right 0.5s ease-out',
-        'slide-in-left': 'slide-in-left 0.5s ease-out',
-        'bounce-in': 'bounce-in 0.8s cubic-bezier(0.68, -0.55, 0.265, 1.55)',
-        'glow': 'glow 2s ease-in-out infinite alternate'
-      }
-    }
-  },
-  plugins: [
-    require("tailwindcss-animate"),
-    require("tailwindcss-rtl"),
-    plugin(function ({ addBase }) {
-      addBase({
-        '@font-face': [
-          {
-            fontFamily: 'Tajawal',
-            fontWeight: '400',
-            fontStyle: 'normal',
-            fontDisplay: 'swap',
-            src: `url('/fonts/Tajawal/Tajawal-Regular.ttf') format('truetype')`,
-          },
-          {
-            fontFamily: 'Amiri',
-            fontWeight: '400',
-            fontStyle: 'normal',
-            fontDisplay: 'swap',
-            src: `url('/fonts/Amiri/Amiri-Regular.ttf') format('truetype')`,
-          },
-          {
-            fontFamily: 'Lalezar',
-            fontWeight: '400',
-            fontStyle: 'normal',
-            fontDisplay: 'swap',
-            src: `url('/fonts/Lalezar/Lalezar-Regular.ttf') format('truetype')`,
-          },
-          {
-            fontFamily: 'Tharwat',
-            fontWeight: '400',
-            fontStyle: 'normal',
-            fontDisplay: 'swap',
-            src: `url('/fonts/TharwatOmaraa.ttf') format('truetype')`,
-          },
-        ],
-      });
-    })
-  ],
-};
