--- frontend/backup-ui/styles/theme.css
@@ -1,57 +1,0 @@
-@layer base {
-  :root {
-    --background: 0 0% 100%;
-    --foreground: 224 71% 4%;
-
-    --primary: 221 83% 53%;
-    --primary-foreground: 0 0% 100%;
-
-    --secondary: 247 82% 70%;
-    --secondary-foreground: 0 0% 100%;
-
-    --accent: 43 96% 56%;
-    --accent-foreground: 24 100% 10%;
-
-    --muted: 210 40% 96%;
-    --muted-foreground: 215 20% 35%;
-
-    --card: 0 0% 100%;
-    --card-foreground: 224 71% 4%;
-
-    --border: 214 32% 91%;
-    --input: 214 32% 91%;
-    --ring: 221 83% 53%;
-
-    --popover: 0 0% 100%;
-    --popover-foreground: 224 71% 4%;
-
-    --radius: 0.5rem;
-  }
-
-  .dark {
-    --background: 224 71% 4%;
-    --foreground: 210 40% 98%;
-
-    --card: 224 71% 6%;
-    --card-foreground: 0 0% 100%;
-
-    --muted: 224 71% 8%;
-    --muted-foreground: 215 20% 65%;
-
-    --secondary: 247 82% 80%;
-    --secondary-foreground: 224 71% 4%;
-
-    --primary: 221 83% 70%;
-    --primary-foreground: 224 71% 4%;
-
-    --accent: 43 96% 64%;
-    --accent-foreground: 224 71% 4%;
-
-    --border: 224 71% 15%;
-    --input: 224 71% 15%;
-    --ring: 221 83% 70%;
-
-    --popover: 224 71% 6%;
-    --popover-foreground: 210 40% 98%;
-  }
-}
