--- frontend/backup-ui/theme.js
@@ -1,11 +1,0 @@
-export const theme = {
-  colors: {
-    primary: 'hsl(var(--primary))',
-    secondary: 'hsl(var(--secondary))',
-    accent: 'hsl(var(--accent))',
-    background: 'hsl(var(--background))',
-    foreground: 'hsl(var(--foreground))',
-  },
-};
-
-export default theme;
